Moved image schema and datamodel files to match jwst structure. Adjusted code paths as neccesary.
