+++ romancal/datamodels/schemas/image.schema.yaml
@@ -1,9 +1,9 @@
 %YAML 1.1
 ---
 $schema: http://stsci.edu/schemas/asdf/asdf-schema-1.0.0
-id: "http://stsci.edu/schemas/roman_datamodel/image_files/image.schema"
+id: "http://stsci.edu/schemas/roman_datamodel/image.schema"
 allOf:
-- $ref: ../core.schema
+- $ref: core.schema
 # To be implemented
 # - $ref: variance.schema
 - type: object
